+++ packages/components/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - **[EXPERIMENTAL]** `getInitialStyles` now accepts a `skeletonTagNames` array of component names that will initially have skeleton styles while the Porsche Design System is loading
 - **[EXPERIMENTAL]** `Button`, `Button Pure`, `Checkbox Wrapper`, `Fieldset Wrapper`, `Headline`, `Link`, `Link Pure`, `Link Social`, `Radio Button Wrapper`, `Select Wrapper`, `Text Field Wrapper`, `Text List`, `Text List Item`, `Textarea Wrapper` can now have skeleton styles based on their default attributes
+- Styles for slotted `<button>` in `Text`
 
 #### Changed
 - `Modal` heading and aria validation happens only when open